pyproject.toml
[project]
name = "base-knowledge"
version = "0.1.0"
description = "base-knowledge"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.4",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "docling>=2.8.1",
    "docling-core>=2.6.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",
    "psycopg2-binary>=2.9.9",
    "langchain>=0.3.17",
    "langchain-openai>=0.3.2",
    "langgraph>=0.2.62",
    "langchain-community>=0.3.16",
    "langchain-core>=0.3.28",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "duckduckgo-search>=8.1.1",
    "duckduckgo-mcp-server>=0.1.0",
    "gradio>=6.1.0",
]
